editoast: extract track occupancy utils and simplify train_schedule
